pipelines/requirements.txt
apache-beam==2.49.0
python-dotenv
pandas
requests